Fixed file loacations
